+++ src/components/icons/index.js
@@ -1,0 +1,2 @@
+export { default as alignLeft } from './align-left';
+export { default as widthInherit } from './width-inherit';