+++ changelog.md
@@ -1,5 +1,11 @@
 # Change log
 
+## [[2.0.0]](https://github.com/lightspeeddevelopment/lsx-mega-menus/releases/tag/2.0.0) - 2023-06-05
+- The LSX Mega Menu has been updated to work with the site editor and the core navigation block.
+
+### Security
+- General testing to ensure compatibility with latest WordPress version (6.2).
+
 ## [[1.2.7]](https://github.com/lightspeeddevelopment/lsx-mega-menus/releases/tag/1.2.7) - 2022-04-20
 
 ### Updated